refactor(web): extract project member dialog content
